fix: constants, db name, signout route, request logger init, routes index rename
