+++ app.js
@@ -6,7 +6,7 @@ const { errors } = require('celebrate');
 const cookieParser = require('cookie-parser');
 const rateLimit = require('express-rate-limit');
 const helmet = require('helmet');
-const router = require('./routes/routes');
+const router = require('./routes/index');
 const {
   PORT, DB_URL, corsConfig, limiterConfig,
 } = require('./config');
@@ -17,6 +17,7 @@ const { requestLogger, errorLogger } = require('./middlewares/logger');
 
 // app init
 const app = express();
+app.use(requestLogger);
 app.use(rateLimit(limiterConfig));
 app.use(helmet());
 app.use(cors(corsConfig));
@@ -26,8 +27,7 @@ app.use(cookieParser());
 // mongodb
 mongoose.connect(DB_URL, { useNewUrlParser: true });
 
-// routes & loggers
-app.use(requestLogger);
+// routes & error logger
 app.use('/', router);
 app.use(errorLogger);
