+++ config.js
@@ -18,7 +18,7 @@ const limiterConfig = {
 
 module.exports = {
   JWT_SECRET: NODE_ENV === 'production' ? JWT_SECRET : 'dev_secret_key',
-  DB_URL: NODE_ENV === 'production' ? DB_URL : 'mongodb://localhost:27017/moviesdb',
+  DB_URL: NODE_ENV === 'production' ? DB_URL : 'mongodb://localhost:27017/bitfilmsdb',
   PORT: NODE_ENV === 'production' ? PORT : 3000,
   corsConfig,
   limiterConfig,
