+++ routes/users.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const { celebrate, Joi } = require('celebrate');
-const { getUserById, updateUserInfo } = require('../controllers/users');
+const { getUserById, updateUserInfo, logout } = require('../controllers/users');
 
 // get current user info
 router.get('/me', getUserById);
@@ -12,5 +12,8 @@ router.patch('/me', celebrate({
     name: Joi.string().min(2).max(30).required(),
   }),
 }), updateUserInfo);
+
+// sign out
+router.post('/signout', logout);
 
 module.exports = router;
